+++ persistence/src/main/java/cz/siemens/inventory/entity/ApplianceRevision.java
@@ -28,7 +28,7 @@ public class ApplianceRevision implements Serializable {
 	private Long id;
 	
 	@Column(name = "rev_interval")
-	private byte interval;
+	private int interval;
 	
 	@Column(name = "last_revision")
 	private LocalDate lastRevision;
